Improve `LoginFlow`.
