+++ client/src/main/kotlin/io/spine/examples/pingh/client/LoginFailed.kt
@@ -1,0 +1,60 @@
+/*
+ * Copyright 2024, TeamDev. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.spine.examples.pingh.client
+
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+
+/**
+ * A stage in the login flow that indicates a failure in the process.
+ *
+ * Possible reasons for failure include:
+ *
+ * 1. The user is not a member of an authorized organization.
+ * 2. The username obtained in [EnterUsername] stage differs from the username
+ * of the account used to complete [VerifyLogin] stage.
+ *
+ * @property moveToNextStage Switches the current stage to the [EnterUsername].
+ * @param cause The reason for the login failure.
+ */
+public class LoginFailed internal constructor(
+    private val moveToNextStage: () -> Unit,
+    cause: String
+) : LoginStage<Unit>() {
+
+    /**
+     * The error message from the login process.
+     */
+    public val errorMessage: StateFlow<String> = MutableStateFlow(cause)
+
+    /**
+     * Starts the login process from the beginning.
+     */
+    public fun restartLogin() {
+        moveToNextStage()
+    }
+}
